refactor(pages): split app.rs (1441 lines) into submodules
Move responsibilities out of the monolithic `crates/reinhardt-pages/src/app.rs`
into focused submodules under `app/`, keeping the module root as a thin
orchestrator that declares submodules, retains the `spa_router` helpers
(`with_router`, `store_spa_router`), and re-exports the public surface so
no downstream code path changes.

Layout after the split (Rust 2024 Edition; no `mod.rs`):

  crates/reinhardt-pages/src/
  ├── app.rs                       (~109 lines: submodule decls, helpers, re-exports)
  └── app/
      ├── launcher.rs              (~1098 lines: ClientLauncher + LaunchCtx + PathCtx
      │                             + PathSubscription + impl blocks + their tests)
      ├── link_interceptor.rs      (~266 lines: AnchorAttrs + should_intercept +
      │                             install_link_interceptor + their tests)
      └── spa_router.rs            (unchanged)

Internal visibility was tightened to the most restrictive that still
compiles (`pub(super)` / `pub(crate)`); public API is preserved via
`pub use` re-exports in `app.rs`.

Verification:
- `cargo fmt -p reinhardt-pages -- --check`
- `cargo check -p reinhardt-pages --all-features`
- `cargo clippy -p reinhardt-pages --all-features -- -D warnings`
- `cargo nextest run -p reinhardt-pages --all-features`

Fixes #4314
